+++ doc/changes.rst
@@ -10,7 +10,9 @@ Change Log
 1.9.15 (unreleased)
 -------------------
 
-* No changes yet.
+* Set read-only permissions on all Module files, and unlock them as needed (PR `#125`_).
+
+.. _`#125`: https://github.com/desihub/desiutil/pull/125
 
 1.9.14 (2018-10-05)
 -------------------