+++ doc/changes.rst
@@ -10,7 +10,11 @@ Change Log
 1.9.15 (unreleased)
 -------------------
 
-* Draw ecliptic in all-sky plots.
+* Set read-only permissions on all Module files, and unlock them as needed (PR `#125`_).
+* Draw ecliptic in all-sky plots (PR `#124`_).
+
+.. _`#125`: https://github.com/desihub/desiutil/pull/125
+.. _`#124`: https://github.com/desihub/desiutil/pull/124
 
 1.9.14 (2018-10-05)
 -------------------